// src/types/offer.ts
import { User } from './user';

export type Location = {
  latitude: number;
  longitude: number;
  zoom: number;
}

export type City = {
  name: string;
  location: Location;
}

export type Offer = {
  id: string;
  title: string;
  type: string;
  price: number;
  city: City;
  location: Location;
  isFavorite: boolean;
  isPremium: boolean;
  rating: number;
  previewImage?: string;
  description?: string;
  bedrooms?: number;
  goods?: string[];
  host?: User;
  images?: string[];
  maxAdults?: number;
}